// src/api/generated/model/getUserDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */

export interface GetUserDto {
  /**
   * Unique identifier for the user
   * @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
   */
  id: string;
  /**
   * Email address of the user
   * @pattern ^(?!\.)(?!.*\.\.)([A-Za-z0-9_'+\-\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\-]*\.)+[A-Za-z]{2,}$
   */
  email: string;
  /** Name of the user */
  name: string;
  /** CPF of the user */
  cpf: string;
  avatarUrl?: string;
  bio?: string;
  /** Date when the user was created */
  createdAt: string;
  /** Date when the user was last updated */
  updatedAt: string;
}