// src/api/generated/model/getPaginatedPostDtoDataItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { _GetPaginatedPostDtoDataItemCount } from "./_getPaginatedPostDtoDataItemCount";
import type { GetPaginatedPostDtoDataItemMediaItem } from "./getPaginatedPostDtoDataItemMediaItem";
import type { GetPaginatedPostDtoDataItemUser } from "./getPaginatedPostDtoDataItemUser";

export type GetPaginatedPostDtoDataItem = {
  /**
   * Identificador único do post
   * @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
   */
  id: string;
  /** Legenda do post */
  caption: string;
  /** Define se o post é público */
  public: boolean;
  /** Data de criação do post */
  createdAt: string;
  /** Data da última atualização do post */
  updatedAt: string;
  /**
   * Identificador do usuário que criou o post
   * @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
   */
  userId: string;
  /** Indica se o post é um vídeo */
  isVideo: boolean;
  /** Lista de mídias (imagens/vídeos) do post */
  Media?: GetPaginatedPostDtoDataItemMediaItem[];
  /** URL da imagem/vídeo do media */
  thumbnailUrl: string;
  _count: _GetPaginatedPostDtoDataItemCount;
  user: GetPaginatedPostDtoDataItemUser;
};